+++ ChangeLog.md
@@ -57,4 +57,5 @@
    * Castelia City cafe guy will use correct types when referring to Striaton City leaders, or use generic terms when completely random
    * Castelia City Gym Trainers and Burgh will use the true type of the gym when type-theming, or use generic terms when completely random.
    * Cheren will use Burgh's true type when referring to him, or use generic terms when completely random.
+   * Nimbasa City Gym Guy will use true type and appropriate weakness when using type-theming on Gym Leaders, or use generic terms when completely random
  * Added `Normalize` and `Stall` to list of negative abilities that can be banned